validation updated

+++ src/main/java/VApp/VApp/entity/BankAccount.java
@@ -1,10 +1,12 @@
 package VApp.VApp.entity;
 
 import jakarta.persistence.*;
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 
-import java.lang.reflect.Type;
-import java.util.UUID;
+import java.math.BigDecimal;
 
 @Entity
 @Getter
@@ -16,15 +18,15 @@ public class BankAccount {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private int id;
-    private String name ;
-    private double balance;
+    private String name;
+    private BigDecimal balance;
     private Long accountNumber;
 
 
     public BankAccount(Account account) {
-        this.id=account.getId();
+        this.id = account.getId();
         this.accountNumber = account.getAccountNumber();
-        this.name=account.getFullName();
-        this.balance=account.getBalance();
+        this.name = account.getFullName();
+        this.balance = account.getBalance();
     }
 }
